tasks 4.1 and 4.4 refactored
