+++ src/main/java/collections/OrdersHandler.java
@@ -1,4 +1,4 @@
-package algorithms;
+package collections;
 
 import java.util.HashSet;
 import java.util.Iterator;
