+++ src/test/java/TestFrequencyCounter.java
@@ -1,22 +1,22 @@
 import org.junit.jupiter.api.Test;
-import algorithms.FrequencyCounter;
+import collections.FrequencyCounter;
 
-import java.util.HashMap;
+import java.util.TreeMap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestFrequencyCounter {
     @Test
     public void testCountCharFrequency() {
-        String testTtext = "abraCadabra";
-        HashMap<String, Integer> testFrequencyDictionary = new HashMap<String, Integer>() {{
+        String testText = "abraCadabra";
+        TreeMap<String, Integer> testFrequencyDictionary = new TreeMap<String, Integer>() {{
+            put("C", 1);
             put("a", 5);
             put("b", 2);
-            put("r", 2);
-            put("C", 1);
             put("d", 1);
+            put("r", 2);
         }};
 
-        assertEquals(testFrequencyDictionary, FrequencyCounter.countCharFrequency(testTtext), "chars counted incorrectly");
+        assertEquals(testFrequencyDictionary, FrequencyCounter.countCharFrequency(testText), "chars counted incorrectly");
     }
 }
